move task manager composable to tasks
