+++ wls-gui-wahllokalsystem/src/composables/tasks/taskManager.ts
